Segunda Aula

+++ demo/src/main/java/negocios/EventoServico.java
@@ -8,9 +8,13 @@ import com.example.demo.entidades.Evento;
 
 @Service
 public interface EventoServico {
-    List<Evento> obterTodosEventos();
-    Evento obterEventoPorId(Long idEvento);
-    void adicionarEvento(Evento evento);
-    void atualizarEvento(Evento evento);
-    void excluirEvento(Long idEvento);
+	List<Evento> obterTodosEventos();
+
+	Evento obterEventoPorId(int idEvento);
+
+	void adicionarEvento(Evento evento);
+
+	void atualizarEvento(Evento evento);
+
+	void excluirEvento(int idEvento);
 }
